New ways to gain experience (#35)
This is the first working build on voice experience. It still needs to
be improved on and tested. Most importantly the experience rate and
multiplier functions will probably need to be altered in the future.

In the future I'm going to have to put together a means of storing
valid games in the database instead of hardcoded IDs or Names,
so plugins like automod and levels can share the same list of valid
games.