Refactor project, introduce readiness probe
Effectively moved the `runner.py` script into `kiki/__main__.py`,
allowing Kiki commands to be executed on the module instead of the
external script. Some references were fixed for this.

Additionally, a readiness probe has been introduced to the Kubernetes
deployment. When Kiki is completely ready, it will write to a health
file that Kubernetes reads periodically, letting Kubernetes know that
Kiki is fully ready and the previous deployment can be torn down with
no downtime.