Run Kiki alongside Starlette
This will allow external services to request Kiki to make changes to a
Discord server without having to interact with Kiki using text based
commands in Discord. For example an administrator will be able to manage
supported games for a server through a web interface, and the service
will tell Kiki to create the necessary roles and emojis for the game.